{"types":"C","runs":12}

+++ MidTerm191.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class MidTerm191 extends ConsoleProgram{
 	public void run(){
-		double a = 12/5;
+		double a =(double) 12/5;
 		print(a);
 	}
 }
